Seal the `Connection` trait
