+++ lib/Cargo.toml
@@ -18,8 +18,8 @@ resolver = "2"
 [features]
 # Public features
 default = ["protocol-ws", "rustls"]
-protocol-http = ["dep:reqwest", "dep:tokio-util", "dep:indexmap"]
-protocol-ws = ["dep:tokio-tungstenite", "dep:tokio-stream", "dep:indexmap", "tokio/time"]
+protocol-http = ["dep:reqwest", "dep:tokio-util"]
+protocol-ws = ["dep:tokio-tungstenite", "dep:tokio-stream", "tokio/time"]
 kv-mem = ["dep:echodb", "parallel"]
 kv-indxdb = ["dep:indxdb"]
 kv-rocksdb = ["dep:rocksdb", "parallel"]
@@ -62,7 +62,7 @@ futures-concurrency = "7.0.0"
 foundationdb = { version = "0.7.0", default-features = false, features = ["embedded-fdb-include"], optional = true }
 fuzzy-matcher = "0.3.7"
 geo = { version = "0.23.0", features = ["use-serde"] }
-indexmap = { version = "1.9.2", optional = true, features = ["serde"] }
+indexmap = { version = "1.9.2", features = ["serde"] }
 indxdb = { version = "0.2.0", optional = true }
 js = { version = "0.1.7", package = "rquickjs", features = ["bindgen", "classes", "futures", "loader", "macro", "properties", "parallel"], optional = true }
 lexical-sort = "0.3.1"
