+++ lib/src/api/engines/local/native.rs
@@ -1,15 +1,15 @@
+use crate::api::conn::Connection;
+use crate::api::conn::DbResponse;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Route;
+use crate::api::conn::Router;
 use crate::api::engines::local::Db;
 use crate::api::opt::from_value;
-use crate::api::opt::DbResponse;
-use crate::api::opt::Param;
 use crate::api::opt::ServerAddrs;
-use crate::api::Connection;
 use crate::api::ExtraFeatures;
-use crate::api::Method;
 use crate::api::QueryResponse;
 use crate::api::Result;
-use crate::api::Route;
-use crate::api::Router;
 use crate::api::Surreal;
 use crate::Datastore;
 use crate::Session;
@@ -25,6 +25,8 @@ use std::marker::PhantomData;
 use std::pin::Pin;
 use std::sync::atomic::AtomicI64;
 use std::sync::Arc;
+
+impl crate::api::Connection for Db {}
 
 impl Connection for Db {
 	fn new(method: Method) -> Self {
