+++ lib/src/api/engines/local/wasm.rs
@@ -1,15 +1,15 @@
 use super::LOG;
+use crate::api::conn::Connection;
+use crate::api::conn::DbResponse;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Route;
+use crate::api::conn::Router;
 use crate::api::engines::local::Db;
 use crate::api::opt::from_value;
-use crate::api::opt::DbResponse;
-use crate::api::opt::Param;
 use crate::api::opt::ServerAddrs;
-use crate::api::Connection;
-use crate::api::Method;
 use crate::api::QueryResponse;
 use crate::api::Result;
-use crate::api::Route;
-use crate::api::Router;
 use crate::api::Surreal;
 use crate::Datastore;
 use crate::Session;
@@ -26,6 +26,8 @@ use std::pin::Pin;
 use std::sync::atomic::AtomicI64;
 use std::sync::Arc;
 use wasm_bindgen_futures::spawn_local;
+
+impl crate::api::Connection for Db {}
 
 impl Connection for Db {
 	fn new(method: Method) -> Self {
