+++ lib/src/api/engines/remote/ws/wasm.rs
@@ -1,21 +1,21 @@
 use super::LOG;
 use super::PATH;
+use crate::api::conn::Connection;
+use crate::api::conn::DbResponse;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Route;
+use crate::api::conn::Router;
 use crate::api::engines::remote::ws::Client;
 use crate::api::engines::remote::ws::Response;
 use crate::api::engines::remote::ws::PING_INTERVAL;
 use crate::api::engines::remote::ws::PING_METHOD;
 use crate::api::err::Error;
 use crate::api::opt::from_value;
-use crate::api::opt::DbResponse;
-use crate::api::opt::Param;
 use crate::api::opt::ServerAddrs;
-use crate::api::Connection;
 use crate::api::ExtraFeatures;
-use crate::api::Method;
 use crate::api::QueryResponse;
 use crate::api::Result;
-use crate::api::Route;
-use crate::api::Router;
 use crate::api::Surreal;
 use crate::sql::Strand;
 use crate::sql::Value;
@@ -54,6 +54,8 @@ pub(crate) enum Either {
 	Event(WsEvent),
 	Ping,
 }
+
+impl crate::api::Connection for Client {}
 
 impl Connection for Client {
 	fn new(method: Method) -> Self {
