+++ lib/src/api/method/authenticate.rs
@@ -1,11 +1,11 @@
-use crate::api::method::Method;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Router;
 use crate::api::opt::auth::Jwt;
-use crate::api::opt::Param;
 use crate::api::Connection;
 use crate::api::Error;
 use crate::api::ExtraFeatures;
 use crate::api::Result;
-use crate::api::Router;
 use std::future::Future;
 use std::future::IntoFuture;
 use std::pin::Pin;
