+++ lib/src/api/method/merge.rs
@@ -1,11 +1,11 @@
-use crate::api::method::Method;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Router;
 use crate::api::opt::from_json;
 use crate::api::opt::DbResource;
-use crate::api::opt::Param;
 use crate::api::opt::Range;
 use crate::api::Connection;
 use crate::api::Result;
-use crate::api::Router;
 use crate::sql::Id;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
