+++ lib/src/api/method/mod.rs
@@ -62,6 +62,7 @@ pub use use_ns::UseNs;
 pub use use_ns::UseNsDb;
 pub use version::Version;
 
+use crate::api::conn::Method;
 use crate::api::opt;
 use crate::api::opt::auth;
 use crate::api::opt::auth::Credentials;
@@ -79,54 +80,6 @@ use serde::Serialize;
 use serde_json::json;
 use std::marker::PhantomData;
 use std::path::Path;
-
-/// The query method
-#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum Method {
-	/// Sends an authentication token to the server
-	Authenticate,
-	/// Perfoms a merge update operation
-	Merge,
-	/// Creates a record in a table
-	Create,
-	/// Deletes a record from a table
-	Delete,
-	/// Exports a database
-	Export,
-	/// Checks the health of the server
-	Health,
-	/// Imports a database
-	Import,
-	/// Invalidates a session
-	Invalidate,
-	/// Kills a live query
-	#[doc(hidden)] // Not supported yet
-	Kill,
-	/// Starts a live query
-	#[doc(hidden)] // Not supported yet
-	Live,
-	/// Perfoms a patch update operation
-	Patch,
-	/// Sends a raw query to the database
-	Query,
-	/// Selects a record or records from a table
-	Select,
-	/// Sets a parameter on the connection
-	Set,
-	/// Signs into the server
-	Signin,
-	/// Signs up on the server
-	Signup,
-	/// Removes a parameter from a connection
-	Unset,
-	/// Perfoms an update operation
-	Update,
-	/// Selects a namespace and database to use
-	Use,
-	/// Queries the version of the server
-	Version,
-}
 
 impl Method {
 	#[allow(dead_code)] // used by `ws` and `http`
@@ -278,7 +231,7 @@ where
 	///
 	/// ```no_run
 	/// use surrealdb::Surreal;
-	/// use surrealdb::engines::remote::{Ws, Wss};
+	/// use surrealdb::engines::remote::ws::{Ws, Wss};
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
