+++ lib/src/api/method/select.rs
@@ -1,10 +1,10 @@
-use crate::api::method::Method;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Router;
 use crate::api::opt::DbResource;
-use crate::api::opt::Param;
 use crate::api::opt::Range;
 use crate::api::Connection;
 use crate::api::Result;
-use crate::api::Router;
 use crate::sql::Id;
 use serde::de::DeserializeOwned;
 use std::future::Future;
