+++ lib/src/api/method/set.rs
@@ -1,8 +1,8 @@
-use crate::api::method::Method;
-use crate::api::opt::Param;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Router;
 use crate::api::Connection;
 use crate::api::Result;
-use crate::api::Router;
 use crate::sql::Value;
 use std::future::Future;
 use std::future::IntoFuture;
