+++ lib/src/api/method/tests/protocol.rs
@@ -1,15 +1,15 @@
 use super::server;
+use crate::api::conn::Connection;
+use crate::api::conn::DbResponse;
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Route;
+use crate::api::conn::Router;
 use crate::api::opt::from_value;
-use crate::api::opt::DbResponse;
-use crate::api::opt::Param;
 use crate::api::opt::ServerAddrs;
 use crate::api::opt::ToServerAddrs;
-use crate::api::Connection;
 use crate::api::ExtraFeatures;
-use crate::api::Method;
 use crate::api::Result;
-use crate::api::Route;
-use crate::api::Router;
 use crate::api::Surreal;
 use crate::QueryResponse;
 use flume::Receiver;
@@ -43,6 +43,8 @@ impl ToServerAddrs<Test> for () {
 pub struct Client {
 	method: Method,
 }
+
+impl crate::api::Connection for Client {}
 
 impl Connection for Client {
 	fn new(method: Method) -> Self {
