+++ lib/src/api/method/tests/server.rs
@@ -1,10 +1,10 @@
 use super::types::Credentials;
 use super::types::User;
-use crate::api::method::Method;
+use crate::api::conn::DbResponse;
+use crate::api::conn::Method;
+use crate::api::conn::Route;
 use crate::api::opt::from_json;
 use crate::api::opt::from_value;
-use crate::api::opt::DbResponse;
-use crate::api::Route;
 use crate::sql::Array;
 use crate::sql::Value;
 use crate::QueryResponse;
