+++ lib/src/api/method/update.rs
@@ -1,14 +1,14 @@
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Router;
 use crate::api::method::Content;
 use crate::api::method::Merge;
-use crate::api::method::Method;
 use crate::api::method::Patch;
 use crate::api::opt::DbResource;
-use crate::api::opt::Param;
 use crate::api::opt::PatchOp;
 use crate::api::opt::Range;
 use crate::api::Connection;
 use crate::api::Result;
-use crate::api::Router;
 use crate::sql::Id;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
