+++ lib/src/api/method/use_db.rs
@@ -1,10 +1,10 @@
-use crate::api::method::Method;
+use crate::api::conn::Method;
 use std::future::Future;
 use std::pin::Pin;
-use crate::api::opt::Param;
+use crate::api::conn::Param;
 use crate::api::Connection;
 use crate::api::Result;
-use crate::api::Router;
+use crate::api::conn::Router;
 use std::future::IntoFuture;
 use crate::sql::Value;
