+++ lib/src/api/method/version.rs
@@ -1,9 +1,9 @@
+use crate::api::conn::Method;
+use crate::api::conn::Param;
+use crate::api::conn::Router;
 use crate::api::err::Error;
-use crate::api::method::Method;
-use crate::api::opt::Param;
 use crate::api::Connection;
 use crate::api::Result;
-use crate::api::Router;
 use std::future::Future;
 use std::future::IntoFuture;
 use std::pin::Pin;
